Extraídas funciones de generación de queries
Las funciones de generación de queries del modelo de perfil de usuario ahora son funciones auxiliares disponibles para distintos modelos.
